requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
autoflake==1.4
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.4
Django==3.0.5
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest==0.8.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
flake8==3.8.4
idna==2.9
importlib-metadata==1.6.0
isort==5.6.4
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-slugify==4.0.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
slugify==0.0.1
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
